Convert ErrorPage and NotFoundPageContent to vanilla-extact
